requirements.txt
requests
python-dotenv
Flask==3.1.0
click
diskcache
feedparser
pytz
pytest
